sidecar - refactor enqueued_telemetry_data and enqueued_telemetry_stats to separate files
