+++ sidecar/src/service/telemetry/mod.rs
@@ -1,2 +1,7 @@
+// Copyright 2021-Present Datadog, Inc. https://www.datadoghq.com/
+// SPDX-License-Identifier: Apache-2.0
+
 pub use app_instance::AppInstance;
 mod app_instance;
+pub mod enqueued_telemetry_data;
+pub mod enqueued_telemetry_stats;
